Extract ProgressLessonLinks
