+++ src/pages/progress/components/ProgressSummaryAndLinks.tsx
@@ -1,7 +1,6 @@
 import React from "react";
-import PARAMS from "../../../utils/params.js";
-import { Link } from "react-router-dom";
 import ProgressSummaryCompleted from "./ProgressSummaryCompleted";
+import ProgressLessonLinks from "./ProgressLessonLinks";
 
 import type { MetWords } from "../../../types";
 
@@ -14,76 +13,6 @@ type Props = {
   ) => void;
   yourMemorisedWordCount: number;
   yourSeenWordCount: number;
-};
-
-type LessonLinkProps = {
-  yourWordCount: number;
-  yourMemorisedWordCount: number;
-  yourSeenWordCount: number;
-};
-
-const getPracticeAllLink = (seen: number, memorised: number) =>
-  seen > 1 && memorised > 0 ? (
-    <>
-      <Link to="/lessons/progress/">Practice&nbsp;your words</Link>.{" "}
-    </>
-  ) : null;
-
-const getDrillMemorisedLink = (memorised: number) =>
-  memorised > 0 ? (
-    <>
-      <Link to="/lessons/progress/memorised/">
-        Drill&nbsp;{memorised} memorised word
-      </Link>
-      .{" "}
-    </>
-  ) : null;
-
-const getReviseSeenLink = (seen: number) =>
-  seen > 0 ? (
-    <>
-      <Link to="/lessons/progress/seen/">Revise&nbsp;{seen} seen words</Link>.{" "}
-    </>
-  ) : null;
-
-const getDiscoverNewLink = (yourWordCount: number) =>
-  yourWordCount < 10000 ? (
-    <>
-      <Link
-        to={
-          "/lessons/drills/top-10000-project-gutenberg-words/?recommended=true&" +
-          PARAMS.discoverParams
-        }
-      >
-        Discover new words
-      </Link>
-      .
-    </>
-  ) : null;
-
-const ProgressLessonLinks = ({
-  yourWordCount,
-  yourSeenWordCount,
-  yourMemorisedWordCount,
-}: LessonLinkProps) => {
-  return (
-    <>
-      {[
-        <React.Fragment key="practice-all">
-          {getPracticeAllLink(yourSeenWordCount, yourMemorisedWordCount)}
-        </React.Fragment>,
-        <React.Fragment key="drill-memorised">
-          {getDrillMemorisedLink(yourMemorisedWordCount)}
-        </React.Fragment>,
-        <React.Fragment key="revise-seen">
-          {getReviseSeenLink(yourSeenWordCount)}
-        </React.Fragment>,
-        <React.Fragment key="discover-new">
-          {getDiscoverNewLink(yourWordCount)}
-        </React.Fragment>,
-      ]}
-    </>
-  );
 };
 
 const ProgressSummaryAndLinks = ({
@@ -139,6 +68,7 @@ const ProgressSummaryAndLinks = ({
     );
   }
   if (yourSeenWordCount === 1 && yourMemorisedWordCount > 1) {
+    // skipped, only pluralisation differs in links
     progressSummaryAndLinks = (
       <p>
         You’ve successfully typed {yourWordCount} words without misstrokes.{" "}
@@ -163,6 +93,7 @@ const ProgressSummaryAndLinks = ({
     );
   }
   if (yourSeenWordCount === 0 && yourMemorisedWordCount > 1) {
+    // skipped, only pluralisation differs in links
     progressSummaryAndLinks = (
       <p>
         You’ve successfully typed {yourWordCount} words without misstrokes.{" "}
